+++ Sources/DamagingEvents.cpp
@@ -1,7 +1,7 @@
 #include "DamagingEvents.h"
 
-float DamagingEvents::DamageCalculator(float WeaknessCoefficient) const {
+uint32_t DamagingEvents::DamageCalculator(float WeaknessCoefficient) const {
     float AlinacakHasar{0};
     AlinacakHasar += static_cast<float>(NormalDamage) * WeaknessCoefficient;
-    return AlinacakHasar;
+    return static_cast<uint32_t>(AlinacakHasar);
 }
